Use encryptIV() and decryptIV() in Sailor
