+++ lib/amqp.js
@@ -2,7 +2,7 @@ var logging = require('./logging.js');
 var info = logging.info;
 var debug = logging.debug;
 var amqplib = require('amqplib');
-var cipher = require('./cipher.js');
+var encryptor = require('./encryptor.js');
 var Q = require('q');
 var _ = require('lodash');
 
@@ -86,7 +86,7 @@ AMQPConnection.prototype.listenQueue = function listenQueue(queueName, callback)
         }
 
         try {
-            var decryptedContent = cipher.decryptMessageContent(message.content);
+            var decryptedContent = encryptor.decryptMessageContent(message.content);
         } catch (err) {
             info('Error occured while parsing message #%j payload (%s)', message.fields.deliveryTag, err.message);
             return self.reject(message);
@@ -124,7 +124,7 @@ AMQPConnection.prototype.sendToExchange = function sendToExchange(exchangeName, 
 AMQPConnection.prototype.sendData = function sendData(data, headers) {
     var settings = this.settings;
     var options = getOptions(headers);
-    var encryptedData = cipher.encryptMessageContent(data);
+    var encryptedData = encryptor.encryptMessageContent(data);
 
     this.sendToExchange(settings.PUBLISH_MESSAGES_TO, settings.DATA_ROUTING_KEY, encryptedData, options);
 };
@@ -133,7 +133,7 @@ AMQPConnection.prototype.sendError = function sendError(err, headers, originalMe
     var settings = this.settings;
     var options = getOptions(headers);
     var payload = {};
-    payload.error = cipher.encryptMessageContent({
+    payload.error = encryptor.encryptMessageContent({
         name: err.name,
         message: err.message,
         stack: err.stack
