+++ lib/sailor.js
@@ -6,6 +6,7 @@ var info = logging.info;
 var _ = require('lodash');
 var Q = require('q');
 var util = require('util');
+var cipher = require('./cipher.js');
 
 var TIMEOUT = process.env.TIMEOUT || 20 * 60 * 1000; // 20 minutes
 
@@ -116,7 +117,7 @@ Sailor.prototype.processMessage = function processMessage(payload, message) {
         compId: stepInfo.compId,
         function: stepInfo.function,
         start: new Date().getTime(),
-        cid: 1
+        cid: cipher.id
     };
 
     return sailor.componentReader.loadTriggerOrAction(stepInfo.function)
