+++ spec/amqp.spec.js
@@ -17,7 +17,7 @@ describe('AMQP', function () {
 
     var AMQPConnection = require('../lib/amqp.js').AMQPConnection;
     var settings = require('../lib/settings.js').readFrom(envVars);
-    var cipher = require('../lib/cipher.js');
+    var encryptor = require('../lib/encryptor.js');
     var _ = require('lodash');
 
     var message = {
@@ -47,7 +47,7 @@ describe('AMQP', function () {
             mandatory: true,
             clusterId: ''
         },
-        content: cipher.encryptMessageContent({"content": "Message content"})
+        content: encryptor.encryptMessageContent({"content": "Message content"})
     };
 
     it('Should send message to outgoing channel when process data', function () {
@@ -79,7 +79,7 @@ describe('AMQP', function () {
             }
         ]);
 
-        var payload = cipher.decryptMessageContent(publishParameters[2].toString());
+        var payload = encryptor.decryptMessageContent(publishParameters[2].toString());
         expect(payload).toEqual({ content : 'Message content' });
     });
 
@@ -113,8 +113,8 @@ describe('AMQP', function () {
         ]);
 
         var payload = JSON.parse(publishParameters[2].toString());
-        payload.error = cipher.decryptMessageContent(payload.error);
-        payload.errorInput = cipher.decryptMessageContent(payload.errorInput);
+        payload.error = encryptor.decryptMessageContent(payload.error);
+        payload.errorInput = encryptor.decryptMessageContent(payload.errorInput);
 
         expect(payload).toEqual({
             error: {
@@ -143,7 +143,7 @@ describe('AMQP', function () {
 
         var publishParameters = amqp.publishChannel.publish.calls[0].args;
         var payload = JSON.parse(publishParameters[2].toString());
-        payload.error = cipher.decryptMessageContent(payload.error);
+        payload.error = encryptor.decryptMessageContent(payload.error);
 
         expect(payload).toEqual({
             error: {
@@ -170,7 +170,7 @@ describe('AMQP', function () {
 
         var publishParameters = amqp.publishChannel.publish.calls[0].args;
         var payload = JSON.parse(publishParameters[2].toString());
-        payload.error = cipher.decryptMessageContent(payload.error);
+        payload.error = encryptor.decryptMessageContent(payload.error);
 
         expect(payload).toEqual({
             error: {
@@ -223,7 +223,7 @@ describe('AMQP', function () {
             }
         ]);
 
-        var payload = cipher.decryptMessageContent(publishParameters[2].toString());
+        var payload = encryptor.decryptMessageContent(publishParameters[2].toString());
         expect(payload).toEqual({content: 'Message content'});
     });
 
@@ -271,7 +271,7 @@ describe('AMQP', function () {
             }
         ]);
 
-        var payload = cipher.decryptMessageContent(publishParameters[2].toString());
+        var payload = encryptor.decryptMessageContent(publishParameters[2].toString());
         expect(payload).toEqual({content: 'Message content'});
     });
 
@@ -319,8 +319,8 @@ describe('AMQP', function () {
 
         var payload = JSON.parse(publishParameters[2].toString());
         console.log(payload);
-        payload.error = cipher.decryptMessageContent(payload.error);
-        payload.errorInput = cipher.decryptMessageContent(payload.errorInput);
+        payload.error = encryptor.decryptMessageContent(payload.error);
+        payload.errorInput = encryptor.decryptMessageContent(payload.errorInput);
 
         expect(payload.error.message).toEqual('Rebound limit exceeded');
         expect(payload.errorInput).toEqual({content : 'Message content'});
@@ -372,7 +372,7 @@ describe('AMQP', function () {
             expect(clientFunction.callCount).toEqual(1);
             expect(clientFunction.calls[0].args[0]).toEqual({"content": "Message content"});
             expect(clientFunction.calls[0].args[1]).toEqual(message);
-            expect(clientFunction.calls[0].args[1].content).toEqual(cipher.encryptMessageContent({"content": "Message content"}));
+            expect(clientFunction.calls[0].args[1].content).toEqual(encryptor.encryptMessageContent({"content": "Message content"}));
         });
     });
