+++ spec/encryptor.spec.js
@@ -3,7 +3,7 @@ describe('Cipher', function () {
     process.env.MESSAGE_CRYPTO_PASSWORD = 'testCryptoPassword';
     process.env.MESSAGE_CRYPTO_IV = 'iv=any16_symbols';
 
-    var cipher = require('../lib/cipher.js');
+    var cipher = require('../lib/encryptor.js');
 
     it('should encrypt & decrypt strings', function () {
         var content = 'Hello world';
